Fix i5OS testcase errors
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/6088


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@23683 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/sdtfmtts.cpp
@@ -243,7 +243,7 @@ void IntlTestSimpleDateFormatAPI::testAPI(/*char *par*/)
     ParsePosition pp(0);
     object.applyPattern("YYYY'W'wwe");
     UDate udDate = object.parse("2007W014", pp);
-    if ((long)udDate == 0) {
+    if ((double)udDate == 0.0) {
         errln("ERROR: Parsing failed using 'Y' and 'e'");
     }
 }
